Aplicativo teclado macro/Sistemas Digitales/requirements.txt
pyserial
pyautogui
tk
